ODS2: Create/Directory fails
Create/directory fails because ODS-2 directories use
VAR records with a MRS of 512 and RAT of NOSPAN.

Ordinarily, MRS of 512 limits the data in a VAR record
to 510 bytes (2 bytes are used for the record length).

In the case of a directory, MRS includes the record length.

ODS2 enforces the MRS, but did not take this case into
account.

Fixes #8