IMD2DSK: Allow to skip maps, specify a filler byte, and improve cyl counting
This PR adds a second invocation for the -i option to skip sector maps.

Also, it adds the -c option to be used a fill byte value for sectors
stored with error marks or not stored at all -- that helps a lot in
discovery of location of such blocks in the .DSK image file, when
an assessment of the file system damage is necessary.

Finally, it improves cylinder counting for IMD files missing some tracks